+++ dev/validation_mandible_orientation.R
@@ -119,7 +119,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_11_CAF,
   individual_id = "MANDIBLE_11_CAF_CBT",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = FALSE
 )
 
@@ -139,7 +139,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_12_CAF,
   individual_id = "MANDIBLE_12_CAF_CBT",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = FALSE
 )
 
@@ -159,7 +159,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_9_CAF,
   individual_id = "MANDIBLE_9_CAF_CBT",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = FALSE
 )
 
@@ -179,7 +179,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_9_CAF,
   individual_id = "MANDIBLE_9_CAF_CBF",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = FALSE,
   lm9_valid = FALSE
 )
@@ -200,7 +200,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_11_CAT,
   individual_id = "MANDIBLE_11_CAT_CBT",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = TRUE
 )
 
@@ -220,7 +220,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_12_CAT,
   individual_id = "MANDIBLE_12_CAT_CBT",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = TRUE
 )
 
@@ -240,7 +240,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_11_CAF,
   individual_id = "MANDIBLE_11_CAF_ELM10",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = FALSE,
   estimate_lm10 = TRUE
 )
@@ -261,7 +261,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_11_CAT,
   individual_id = "MANDIBLE_11_CAT_ELM10",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = TRUE,
   estimate_lm10 = TRUE
 )
@@ -282,7 +282,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_12_CAF,
   individual_id = "MANDIBLE_12_CAF_ELM10",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = FALSE,
   estimate_lm10 = TRUE
 )
@@ -303,7 +303,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_12_CAT,
   individual_id = "MANDIBLE_12_CAT_ELM10",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = TRUE,
   estimate_lm10 = TRUE
 )
@@ -324,7 +324,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_12_CAF,
   individual_id = "MANDIBLE_12_CAF_CBF",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = FALSE,
   lm9_valid = FALSE
 )
@@ -345,7 +345,7 @@ res <- orient_mandible(
   landmarks_str = landmarks_12_CAT,
   individual_id = "MANDIBLE_12_CAT_CBF",
   camera_distance_mm = 300,
-  cs3_camera_side = "RIGHT",
+  lm1_side = "RIGHT",
   complete_arch = TRUE,
   lm9_valid = FALSE
 )